Cleanup Pact tests
